refactor: rename resistor component
