+++ src/ui/components/Result.tsx
@@ -1,7 +1,7 @@
 import { sum, mult } from "correct-operations"
 import { normal_colors, normal_color_name, multiplier_color_name, tolerance_color_name, multiplier_colors, tolerance_colors } from "../../constants"
 import { LineTitle } from "./LineTittle"
-import { colors } from "./Resistor"
+import { colors } from "./resistors/FourBandResistor"
 
 function getBandsValues(colors: colors) {
     const line1_value = normal_colors[colors.line1 as normal_color_name].value
